+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Even Ground - Racial Powers and Birthsigns for NPCs (OpenMW)
 
+## 1.2
+
+- Added new settings
+- Added Fargoth to NPC Whitelist
+- Updated Text Set settings renderer
+- Fixed script initialization ignoring the list of already affected NPCs
+
 ## 1.1.1
 
 - Fixed Actor type filtering